requirements.txt
numpy~=1.24.3
torchvision~=0.18.1
pillow~=10.4.0
matplotlib~=3.7.5
opencv-python~=4.10.0.84